- upgrading to iOS 7/8. - fixed all ? that was either deprecated or changed behavior because of upgrading xcode/ios version. - preparing the ground to start new development on version 2.x

+++ DP/ControllersPaid/DPAnimatedScrollViewController.m
@@ -232,7 +232,7 @@
 }
 - (DPAnimatedCardsView *) findAnimCtgViewInCurrentPage {
     DPAnimatedCardsView *result = nil;
-    int pg = self.pageControl.currentPage;
+    int pg = (int)self.pageControl.currentPage;
     result = [self findAnimCtgViewInPage:pg];
     return result;
 }
